fixed types and file structure
