+++ src/blocks/theme/index.ts
@@ -1,13 +1,4 @@
-export * from './colors/colors.brands';
-export * from './colors/colors.primitives';
-export * from './colors/colors.semantics';
-
 export * from './device/theme.device';
 
 export * from './Theme';
-
-export * from './variables/variables.blur';
-export * from './variables/variables.borderRadius';
-export * from './variables/variables.borderSize';
-export * from './variables/variables.opacity';
-export * from './variables/variables.spacing';
+export * from './Theme.utils';
